backend/requirements.txt
flask
flask-cors
flask_sqlalchemy
bcrypt
pyjwt